Here's what happened
I browsed all branches in the git repository to see what needed to be removed and what needed to be reduced. 

The result is now 2 separate projects that need to be integrated, and a library for interacting with the database.

I am going to commute home now, and will hopefully have the agenda and loginWorking branches merged properly. This will be pushed to the master.

css styling will come later, and I will figure out how to use the database library