Login complete
Finished login system, users can create an account, or log into an existing account. The window prompts errors if the user does something wrong like invalid email, wrong password, or empty field

Found a bug when loading the eventPane, so my next mission is to get the boxes to draw on the dayPane automatically. I created a new class and just commented what it should do in order to do this